+++ spotbugsTestCases/src/java/type/annotation/CheckerFrameworkTypeAnnotations.java
@@ -1,0 +1,13 @@
+package type.annotation;
+
+import org.checkerframework.checker.nullness.qual.NonNull;
+
+public class CheckerFrameworkTypeAnnotations {
+
+    // Expecting NP_NONNULL_RETURN_VIOLATION to be thrown here
+    @NonNull
+    public String foo() {
+        return null;
+    }
+
+}